tree-wide: be more careful when passing literal integers to "t" bus message fields
Since we use varargs for sd_message_append() we need to make sure the
parameters we pass are actually 64bit wide, if "t" is used. Hence cast
appropriately if necessary.

I went through the whole tree, and in most cases we got it right, but
there are some cases we missed so far.

Inspired by: #31420

+++ src/core/dbus-service.c
@@ -61,7 +61,7 @@ static int property_get_open_files(
                 return r;
 
         LIST_FOREACH(open_files, of, *open_files) {
-                r = sd_bus_message_append(reply, "(sst)", of->path, of->fdname, of->flags);
+                r = sd_bus_message_append(reply, "(sst)", of->path, of->fdname, (uint64_t) of->flags);
                 if (r < 0)
                         return r;
         }
